[SIEM] Refactor hosts routing (#47459)
