+++ x-pack/legacy/plugins/siem/public/pages/hosts/details/index.tsx
@@ -11,6 +11,8 @@ import { compose } from 'redux';
 import { connect } from 'react-redux';
 import { StickyContainer } from 'react-sticky';
 
+import { inputsSelectors, State } from '../../../store';
+
 import { FiltersGlobal } from '../../../components/filters_global';
 import { HeaderPage } from '../../../components/header_page';
 import { KpiHostDetailsQuery } from '../../../containers/kpi_host_details';
@@ -32,22 +34,19 @@ import { LastEventIndexKey } from '../../../graphql/types';
 import { convertToBuildEsQuery } from '../../../lib/keury';
 import { setAbsoluteRangeDatePicker as dispatchAbsoluteRangeDatePicker } from '../../../store/inputs/actions';
 import { SpyRoute } from '../../../utils/route/spy_routes';
-
-import { HostsQueryProps } from '../hosts';
-import { HostsEmptyPage } from '../hosts_empty_page';
-
-export { HostDetailsBody } from './body';
-import { navTabsHostDetails } from './nav_tabs';
-import { HostDetailsComponentProps } from './types';
-import { makeMapStateToProps } from './utils';
 import { useKibanaCore } from '../../../lib/compose/kibana_core';
+
+import { HostsEmptyPage } from '../hosts_empty_page';
+import { navTabsHostDetails } from './nav_tabs';
+import { HostDetailsComponentProps, HostDetailsProps } from './types';
+import { HostDetailsTabs } from './details_tabs';
+import { type } from './utils';
 
 const HostOverviewManage = manageQuery(HostOverview);
 const KpiHostDetailsManage = manageQuery(KpiHostsComponent);
 
 const HostDetailsComponent = React.memo<HostDetailsComponentProps>(
   ({
-    detailName,
     filters,
     from,
     isInitializing,
@@ -56,6 +55,9 @@ const HostDetailsComponent = React.memo<HostDetailsComponentProps>(
     setHostDetailsTablesActivePageToZero,
     setQuery,
     to,
+    detailName,
+    deleteQuery,
+    hostDetailsPagePath,
   }) => {
     useEffect(() => {
       setHostDetailsTablesActivePageToZero(null);
@@ -96,94 +98,111 @@ const HostDetailsComponent = React.memo<HostDetailsComponentProps>(
               ],
             });
             return indicesExistOrDataTemporarilyUnavailable(indicesExist) ? (
-              <StickyContainer>
-                <FiltersGlobal>
-                  <SiemSearchBar indexPattern={indexPattern} id="global" />
-                </FiltersGlobal>
+              <>
+                <StickyContainer>
+                  <FiltersGlobal>
+                    <SiemSearchBar indexPattern={indexPattern} id="global" />
+                  </FiltersGlobal>
 
-                <HeaderPage
-                  subtitle={
-                    <LastEventTime indexKey={LastEventIndexKey.hostDetails} hostName={detailName} />
-                  }
-                  title={detailName}
-                />
-                <HostOverviewByNameQuery
-                  sourceId="default"
-                  hostName={detailName}
-                  skip={isInitializing}
-                  startDate={from}
-                  endDate={to}
-                >
-                  {({ hostOverview, loading, id, inspect, refetch }) => (
-                    <AnomalyTableProvider
-                      criteriaFields={hostToCriteria(hostOverview)}
-                      startDate={from}
-                      endDate={to}
-                      skip={isInitializing}
-                    >
-                      {({ isLoadingAnomaliesData, anomaliesData }) => (
-                        <HostOverviewManage
-                          id={id}
-                          inspect={inspect}
-                          refetch={refetch}
-                          setQuery={setQuery}
-                          data={hostOverview}
-                          anomaliesData={anomaliesData}
-                          isLoadingAnomaliesData={isLoadingAnomaliesData}
-                          loading={loading}
-                          startDate={from}
-                          endDate={to}
-                          narrowDateRange={(score, interval) => {
-                            const fromTo = scoreIntervalToDateTime(score, interval);
-                            setAbsoluteRangeDatePicker({
-                              id: 'global',
-                              from: fromTo.from,
-                              to: fromTo.to,
-                            });
-                          }}
-                        />
-                      )}
-                    </AnomalyTableProvider>
-                  )}
-                </HostOverviewByNameQuery>
+                  <HeaderPage
+                    subtitle={
+                      <LastEventTime
+                        indexKey={LastEventIndexKey.hostDetails}
+                        hostName={detailName}
+                      />
+                    }
+                    title={detailName}
+                  />
+                  <HostOverviewByNameQuery
+                    sourceId="default"
+                    hostName={detailName}
+                    skip={isInitializing}
+                    startDate={from}
+                    endDate={to}
+                  >
+                    {({ hostOverview, loading, id, inspect, refetch }) => (
+                      <AnomalyTableProvider
+                        criteriaFields={hostToCriteria(hostOverview)}
+                        startDate={from}
+                        endDate={to}
+                        skip={isInitializing}
+                      >
+                        {({ isLoadingAnomaliesData, anomaliesData }) => (
+                          <HostOverviewManage
+                            id={id}
+                            inspect={inspect}
+                            refetch={refetch}
+                            setQuery={setQuery}
+                            data={hostOverview}
+                            anomaliesData={anomaliesData}
+                            isLoadingAnomaliesData={isLoadingAnomaliesData}
+                            loading={loading}
+                            startDate={from}
+                            endDate={to}
+                            narrowDateRange={(score, interval) => {
+                              const fromTo = scoreIntervalToDateTime(score, interval);
+                              setAbsoluteRangeDatePicker({
+                                id: 'global',
+                                from: fromTo.from,
+                                to: fromTo.to,
+                              });
+                            }}
+                          />
+                        )}
+                      </AnomalyTableProvider>
+                    )}
+                  </HostOverviewByNameQuery>
 
-                <EuiHorizontalRule />
+                  <EuiHorizontalRule />
 
-                <KpiHostDetailsQuery
-                  sourceId="default"
+                  <KpiHostDetailsQuery
+                    sourceId="default"
+                    filterQuery={filterQuery}
+                    skip={isInitializing}
+                    startDate={from}
+                    endDate={to}
+                  >
+                    {({ kpiHostDetails, id, inspect, loading, refetch }) => (
+                      <KpiHostDetailsManage
+                        data={kpiHostDetails}
+                        from={from}
+                        id={id}
+                        inspect={inspect}
+                        loading={loading}
+                        refetch={refetch}
+                        setQuery={setQuery}
+                        to={to}
+                        narrowDateRange={(min: number, max: number) => {
+                          setAbsoluteRangeDatePicker({ id: 'global', from: min, to: max });
+                        }}
+                      />
+                    )}
+                  </KpiHostDetailsQuery>
+
+                  <SiemNavigation
+                    navTabs={navTabsHostDetails(detailName, hasMlUserPermissions(capabilities))}
+                    display="default"
+                    showBorder={true}
+                  />
+                  <EuiSpacer />
+                </StickyContainer>
+                <HostDetailsTabs
+                  isInitializing={isInitializing}
+                  deleteQuery={deleteQuery}
+                  to={to}
+                  from={from}
+                  detailName={detailName}
+                  type={type}
+                  setQuery={setQuery}
                   filterQuery={filterQuery}
-                  skip={isInitializing}
-                  startDate={from}
-                  endDate={to}
-                >
-                  {({ kpiHostDetails, id, inspect, loading, refetch }) => (
-                    <KpiHostDetailsManage
-                      data={kpiHostDetails}
-                      from={from}
-                      id={id}
-                      inspect={inspect}
-                      loading={loading}
-                      refetch={refetch}
-                      setQuery={setQuery}
-                      to={to}
-                      narrowDateRange={(min: number, max: number) => {
-                        setAbsoluteRangeDatePicker({ id: 'global', from: min, to: max });
-                      }}
-                    />
-                  )}
-                </KpiHostDetailsQuery>
-
-                <SiemNavigation
-                  navTabs={navTabsHostDetails(detailName, hasMlUserPermissions(capabilities))}
-                  display="default"
-                  showBorder={true}
+                  hostDetailsPagePath={hostDetailsPagePath}
+                  indexPattern={indexPattern}
+                  setAbsoluteRangeDatePicker={setAbsoluteRangeDatePicker}
                 />
-                <EuiSpacer />
-              </StickyContainer>
+              </>
             ) : (
               <>
                 <HeaderPage title={detailName} />
-
                 <HostsEmptyPage />
               </>
             );
@@ -197,7 +216,16 @@ const HostDetailsComponent = React.memo<HostDetailsComponentProps>(
 
 HostDetailsComponent.displayName = 'HostDetailsComponent';
 
-export const HostDetails = compose<React.ComponentClass<HostsQueryProps & { detailName: string }>>(
+export const makeMapStateToProps = () => {
+  const getGlobalQuerySelector = inputsSelectors.globalQuerySelector();
+  const getGlobalFiltersQuerySelector = inputsSelectors.globalFiltersQuerySelector();
+  return (state: State) => ({
+    query: getGlobalQuerySelector(state),
+    filters: getGlobalFiltersQuerySelector(state),
+  });
+};
+
+export const HostDetails = compose<React.ComponentClass<HostDetailsProps>>(
   connect(
     makeMapStateToProps,
     {
