+++ x-pack/legacy/plugins/siem/public/pages/hosts/details/types.ts
@@ -4,6 +4,7 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { StaticIndexPattern } from 'ui/index_patterns';
 import { Filter } from '@kbn/es-query';
 import { ActionCreator } from 'typescript-fsa';
 import { Query } from 'src/plugins/data/common';
@@ -11,13 +12,10 @@ import { Query } from 'src/plugins/data/common';
 import { InputsModelId } from '../../../store/inputs/constants';
 import { HostComponentProps } from '../../../components/link_to/redirect_to_hosts';
 import { HostsTableType } from '../../../store/hosts/model';
-import { HostsQueryProps } from '../hosts';
+import { HostsQueryProps } from '../types';
 import { NavTab } from '../../../components/navigation/types';
-import {
-  AnomaliesChildren,
-  CommonChildren,
-  KeyHostsNavTabWithoutMlPermission,
-} from '../navigation/types';
+import { KeyHostsNavTabWithoutMlPermission } from '../navigation/types';
+import { hostsModel } from '../../../store';
 
 interface HostDetailsComponentReduxProps {
   query: Query;
@@ -31,24 +29,22 @@ interface HostBodyComponentDispatchProps {
     to: number;
   }>;
   detailName: string;
+  hostDetailsPagePath: string;
 }
 
 interface HostDetailsComponentDispatchProps extends HostBodyComponentDispatchProps {
   setHostDetailsTablesActivePageToZero: ActionCreator<null>;
 }
 
-export interface HostDetailsBodyProps extends HostsQueryProps {
-  children: CommonChildren | AnomaliesChildren;
+export interface HostDetailsProps extends HostsQueryProps {
+  detailName: string;
+  hostDetailsPagePath: string;
 }
 
 export type HostDetailsComponentProps = HostDetailsComponentReduxProps &
   HostDetailsComponentDispatchProps &
   HostComponentProps &
   HostsQueryProps;
-
-export type HostDetailsBodyComponentProps = HostDetailsComponentReduxProps &
-  HostBodyComponentDispatchProps &
-  HostDetailsBodyProps;
 
 type KeyHostDetailsNavTabWithoutMlPermission = HostsTableType.authentications &
   HostsTableType.uncommonProcesses &
@@ -62,3 +58,16 @@ type KeyHostDetailsNavTab =
   | KeyHostDetailsNavTabWithMlPermission;
 
 export type HostDetailsNavTab = Record<KeyHostDetailsNavTab, NavTab>;
+
+export type HostDetailsTabsProps = HostBodyComponentDispatchProps &
+  HostsQueryProps & {
+    indexPattern: StaticIndexPattern;
+    type: hostsModel.HostsType;
+    filterQuery: string;
+  };
+
+export type SetAbsoluteRangeDatePicker = ActionCreator<{
+  id: InputsModelId;
+  from: number;
+  to: number;
+}>;
