+++ x-pack/legacy/plugins/siem/public/pages/hosts/hosts.tsx
@@ -5,13 +5,11 @@
  */
 
 import { EuiSpacer } from '@elastic/eui';
-import { Filter, getEsQueryConfig } from '@kbn/es-query';
+import { getEsQueryConfig } from '@kbn/es-query';
 import * as React from 'react';
 import { compose } from 'redux';
 import { connect } from 'react-redux';
 import { StickyContainer } from 'react-sticky';
-import { ActionCreator } from 'typescript-fsa';
-import { Query } from 'src/plugins/data/common';
 
 import { FiltersGlobal } from '../../components/filters_global';
 import { GlobalTimeArgs } from '../../containers/global_time';
@@ -27,41 +25,34 @@ import { SiemSearchBar } from '../../components/search_bar';
 import { indicesExistOrDataTemporarilyUnavailable, WithSource } from '../../containers/source';
 import { LastEventIndexKey } from '../../graphql/types';
 import { convertToBuildEsQuery } from '../../lib/keury';
-import { inputsSelectors, State } from '../../store';
+import { inputsSelectors, State, hostsModel } from '../../store';
 import { setAbsoluteRangeDatePicker as dispatchSetAbsoluteRangeDatePicker } from '../../store/inputs/actions';
-import { InputsModelId } from '../../store/inputs/constants';
 import { SpyRoute } from '../../utils/route/spy_routes';
+import { useKibanaCore } from '../../lib/compose/kibana_core';
 
 import { HostsEmptyPage } from './hosts_empty_page';
 import { navTabsHosts } from './nav_tabs';
 import * as i18n from './translations';
-import { useKibanaCore } from '../../lib/compose/kibana_core';
+import { HostsComponentProps, HostsComponentReduxProps } from './types';
+import { HostsTabs } from './hosts_tabs';
 
 const KpiHostsComponentManage = manageQuery(KpiHostsComponent);
 
-interface HostsComponentReduxProps {
-  query: Query;
-  filters: Filter[];
-}
-
-interface HostsComponentDispatchProps {
-  setAbsoluteRangeDatePicker: ActionCreator<{
-    id: InputsModelId;
-    from: number;
-    to: number;
-  }>;
-}
-
-export type HostsQueryProps = GlobalTimeArgs;
-
-export type HostsComponentProps = HostsComponentReduxProps &
-  HostsComponentDispatchProps &
-  HostsQueryProps;
-
 const HostsComponent = React.memo<HostsComponentProps>(
-  ({ isInitializing, filters, from, query, setAbsoluteRangeDatePicker, setQuery, to }) => {
+  ({
+    deleteQuery,
+    isInitializing,
+    filters,
+    from,
+    query,
+    setAbsoluteRangeDatePicker,
+    setQuery,
+    to,
+    hostsPagePath,
+  }) => {
     const capabilities = React.useContext(MlCapabilitiesContext);
     const core = useKibanaCore();
+
     return (
       <>
         <WithSource sourceId="default">
@@ -73,48 +64,62 @@ const HostsComponent = React.memo<HostsComponentProps>(
               filters,
             });
             return indicesExistOrDataTemporarilyUnavailable(indicesExist) ? (
-              <StickyContainer>
-                <FiltersGlobal>
-                  <SiemSearchBar indexPattern={indexPattern} id="global" />
-                </FiltersGlobal>
+              <>
+                <StickyContainer>
+                  <FiltersGlobal>
+                    <SiemSearchBar indexPattern={indexPattern} id="global" />
+                  </FiltersGlobal>
 
-                <HeaderPage
-                  subtitle={<LastEventTime indexKey={LastEventIndexKey.hosts} />}
-                  title={i18n.PAGE_TITLE}
-                />
-                <>
-                  <KpiHostsQuery
-                    endDate={to}
-                    filterQuery={filterQuery}
-                    skip={isInitializing}
-                    sourceId="default"
-                    startDate={from}
-                  >
-                    {({ kpiHosts, loading, id, inspect, refetch }) => (
-                      <KpiHostsComponentManage
-                        data={kpiHosts}
-                        from={from}
-                        id={id}
-                        inspect={inspect}
-                        loading={loading}
-                        refetch={refetch}
-                        setQuery={setQuery}
-                        to={to}
-                        narrowDateRange={(min: number, max: number) => {
-                          setAbsoluteRangeDatePicker({ id: 'global', from: min, to: max });
-                        }}
-                      />
-                    )}
-                  </KpiHostsQuery>
-                  <EuiSpacer />
-                  <SiemNavigation
-                    navTabs={navTabsHosts(hasMlUserPermissions(capabilities))}
-                    display="default"
-                    showBorder={true}
+                  <HeaderPage
+                    subtitle={<LastEventTime indexKey={LastEventIndexKey.hosts} />}
+                    title={i18n.PAGE_TITLE}
                   />
-                  <EuiSpacer />
-                </>
-              </StickyContainer>
+                  <>
+                    <KpiHostsQuery
+                      endDate={to}
+                      filterQuery={filterQuery}
+                      skip={isInitializing}
+                      sourceId="default"
+                      startDate={from}
+                    >
+                      {({ kpiHosts, loading, id, inspect, refetch }) => (
+                        <KpiHostsComponentManage
+                          data={kpiHosts}
+                          from={from}
+                          id={id}
+                          inspect={inspect}
+                          loading={loading}
+                          refetch={refetch}
+                          setQuery={setQuery}
+                          to={to}
+                          narrowDateRange={(min: number, max: number) => {
+                            setAbsoluteRangeDatePicker({ id: 'global', from: min, to: max });
+                          }}
+                        />
+                      )}
+                    </KpiHostsQuery>
+                    <EuiSpacer />
+                    <SiemNavigation
+                      navTabs={navTabsHosts(hasMlUserPermissions(capabilities))}
+                      display="default"
+                      showBorder={true}
+                    />
+                    <EuiSpacer />
+                  </>
+                </StickyContainer>
+                <HostsTabs
+                  deleteQuery={deleteQuery}
+                  to={to}
+                  filterQuery={filterQuery}
+                  isInitializing={isInitializing}
+                  setQuery={setQuery}
+                  from={from}
+                  type={hostsModel.HostsType.page}
+                  indexPattern={indexPattern}
+                  setAbsoluteRangeDatePicker={setAbsoluteRangeDatePicker}
+                  hostsPagePath={hostsPagePath}
+                />
+              </>
             ) : (
               <>
                 <HeaderPage title={i18n.PAGE_TITLE} />
@@ -142,8 +147,12 @@ const makeMapStateToProps = () => {
   return mapStateToProps;
 };
 
+interface HostsProps extends GlobalTimeArgs {
+  hostsPagePath: string;
+}
+
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
-export const Hosts = compose<React.ComponentClass<GlobalTimeArgs>>(
+export const Hosts = compose<React.ComponentClass<HostsProps>>(
   connect(
     makeMapStateToProps,
     {
