+++ x-pack/legacy/plugins/siem/public/pages/hosts/index.tsx
@@ -7,21 +7,13 @@
 import React from 'react';
 import { Redirect, Route, Switch, RouteComponentProps } from 'react-router-dom';
 
-import { HostDetailsBody, HostDetails } from './details';
-import {
-  HostsQueryTabBody,
-  AuthenticationsQueryTabBody,
-  UncommonProcessQueryTabBody,
-  AnomaliesQueryTabBody,
-  EventsQueryTabBody,
-} from './navigation';
-import { HostsBody } from './hosts_body';
+import { HostDetails } from './details';
 import { HostsTableType } from '../../store/hosts/model';
+
 import { GlobalTime } from '../../containers/global_time';
 import { SiemPageName } from '../home/types';
 import { Hosts } from './hosts';
-
-const hostsPagePath = `/:pageName(${SiemPageName.hosts})`;
+import { hostsPagePath, hostDetailsPagePath } from './types';
 
 const getHostsTabPath = (pagePath: string) =>
   `${pagePath}/:tabName(` +
@@ -32,7 +24,7 @@ const getHostsTabPath = (pagePath: string) =>
   `${HostsTableType.events})`;
 
 const getHostDetailsTabPath = (pagePath: string) =>
-  `${pagePath}/:detailName/:tabName(` +
+  `${hostDetailsPagePath}/:tabName(` +
   `${HostsTableType.authentications}|` +
   `${HostsTableType.uncommonProcesses}|` +
   `${HostsTableType.anomalies}|` +
@@ -47,188 +39,35 @@ export const HostsContainer = React.memo<Props>(({ url }) => (
         <Route
           strict
           exact
-          path={hostsPagePath}
+          path={getHostsTabPath(hostsPagePath)}
           render={() => (
-            <Route
-              path={hostsPagePath}
-              render={() => (
-                <>
-                  <Hosts from={from} to={to} setQuery={setQuery} isInitializing={isInitializing} />
-                  <HostsBody
-                    deleteQuery={deleteQuery}
-                    from={from}
-                    to={to}
-                    setQuery={setQuery}
-                    isInitializing={isInitializing}
-                    children={HostsQueryTabBody}
-                  />
-                </>
-              )}
+            <Hosts
+              hostsPagePath={hostsPagePath}
+              from={from}
+              to={to}
+              setQuery={setQuery}
+              isInitializing={isInitializing}
+              deleteQuery={deleteQuery}
             />
           )}
         />
         <Route
           strict
-          exact
-          path={getHostsTabPath(hostsPagePath)}
-          render={() => (
-            <>
-              <Hosts from={from} to={to} setQuery={setQuery} isInitializing={isInitializing} />
-              <Route
-                path={`${hostsPagePath}/:tabName(${HostsTableType.hosts})`}
-                render={() => (
-                  <HostsBody
-                    deleteQuery={deleteQuery}
-                    from={from}
-                    to={to}
-                    setQuery={setQuery}
-                    isInitializing={isInitializing}
-                    children={HostsQueryTabBody}
-                  />
-                )}
-              />
-              <Route
-                path={`${hostsPagePath}/:tabName(${HostsTableType.authentications})`}
-                render={() => (
-                  <HostsBody
-                    deleteQuery={deleteQuery}
-                    from={from}
-                    to={to}
-                    setQuery={setQuery}
-                    isInitializing={isInitializing}
-                    children={AuthenticationsQueryTabBody}
-                  />
-                )}
-              />
-              <Route
-                path={`${hostsPagePath}/:tabName(${HostsTableType.uncommonProcesses})`}
-                render={() => (
-                  <HostsBody
-                    deleteQuery={deleteQuery}
-                    from={from}
-                    to={to}
-                    setQuery={setQuery}
-                    isInitializing={isInitializing}
-                    children={UncommonProcessQueryTabBody}
-                  />
-                )}
-              />
-              <Route
-                path={`${hostsPagePath}/:tabName(${HostsTableType.anomalies})`}
-                render={() => (
-                  <HostsBody
-                    deleteQuery={deleteQuery}
-                    from={from}
-                    to={to}
-                    setQuery={setQuery}
-                    isInitializing={isInitializing}
-                    children={AnomaliesQueryTabBody}
-                  />
-                )}
-              />
-              <Route
-                path={`${hostsPagePath}/:tabName(${HostsTableType.events})`}
-                render={() => (
-                  <HostsBody
-                    deleteQuery={deleteQuery}
-                    from={from}
-                    to={to}
-                    setQuery={setQuery}
-                    isInitializing={isInitializing}
-                    children={EventsQueryTabBody}
-                  />
-                )}
-              />
-            </>
-          )}
-        />
-        <Route
-          strict
-          exact
           path={getHostDetailsTabPath(hostsPagePath)}
           render={props => (
-            <>
-              <HostDetails
-                from={from}
-                to={to}
-                setQuery={setQuery}
-                isInitializing={isInitializing}
-                detailName={props.match.params.detailName}
-              />
-              <Route
-                path={`${hostsPagePath}/:detailName/:tabName(${HostsTableType.hosts})`}
-                render={() => (
-                  <HostDetailsBody
-                    deleteQuery={deleteQuery}
-                    from={from}
-                    to={to}
-                    setQuery={setQuery}
-                    isInitializing={isInitializing}
-                    children={HostsQueryTabBody}
-                    detailName={props.match.params.detailName}
-                  />
-                )}
-              />
-              <Route
-                path={`${hostsPagePath}/:detailName/:tabName(${HostsTableType.authentications})`}
-                render={() => (
-                  <HostDetailsBody
-                    deleteQuery={deleteQuery}
-                    from={from}
-                    to={to}
-                    setQuery={setQuery}
-                    isInitializing={isInitializing}
-                    detailName={props.match.params.detailName}
-                    children={AuthenticationsQueryTabBody}
-                  />
-                )}
-              />
-              <Route
-                path={`${hostsPagePath}/:detailName/:tabName(${HostsTableType.uncommonProcesses})`}
-                render={() => (
-                  <HostDetailsBody
-                    deleteQuery={deleteQuery}
-                    from={from}
-                    to={to}
-                    setQuery={setQuery}
-                    isInitializing={isInitializing}
-                    detailName={props.match.params.detailName}
-                    children={UncommonProcessQueryTabBody}
-                  />
-                )}
-              />
-              <Route
-                path={`${hostsPagePath}/:detailName/:tabName(${HostsTableType.anomalies})`}
-                render={() => (
-                  <HostDetailsBody
-                    deleteQuery={deleteQuery}
-                    from={from}
-                    to={to}
-                    setQuery={setQuery}
-                    isInitializing={isInitializing}
-                    detailName={props.match.params.detailName}
-                    children={AnomaliesQueryTabBody}
-                  />
-                )}
-              />
-              <Route
-                path={`${hostsPagePath}/:detailName/:tabName(${HostsTableType.events})`}
-                render={() => (
-                  <HostDetailsBody
-                    from={from}
-                    to={to}
-                    setQuery={setQuery}
-                    isInitializing={isInitializing}
-                    detailName={props.match.params.detailName}
-                    children={EventsQueryTabBody}
-                  />
-                )}
-              />
-            </>
+            <HostDetails
+              hostDetailsPagePath={hostDetailsPagePath}
+              detailName={props.match.params.detailName}
+              from={from}
+              to={to}
+              setQuery={setQuery}
+              isInitializing={isInitializing}
+              deleteQuery={deleteQuery}
+            />
           )}
         />
         <Route
-          path={`${url}/:detailName`}
+          path={hostDetailsPagePath}
           render={({
             match: {
               params: { detailName },
@@ -237,12 +76,9 @@ export const HostsContainer = React.memo<Props>(({ url }) => (
           }) => <Redirect to={`${url}/${detailName}/${HostsTableType.authentications}${search}`} />}
         />
         <Route
-          path={`/${SiemPageName.hosts}/`}
+          path={`${hostsPagePath}/`}
           render={({ location: { search = '' } }) => (
-            <Redirect
-              from={`/${SiemPageName.hosts}/"`}
-              to={`/${SiemPageName.hosts}/${HostsTableType.hosts}${search}`}
-            />
+            <Redirect to={`/${SiemPageName.hosts}/${HostsTableType.hosts}${search}`} />
           )}
         />
       </Switch>
