+++ x-pack/legacy/plugins/siem/public/pages/hosts/types.ts
@@ -1,0 +1,45 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+import { StaticIndexPattern } from 'ui/index_patterns';
+import { ActionCreator } from 'typescript-fsa';
+import { Filter } from '@kbn/es-query';
+import { Query } from 'src/plugins/data/common';
+
+import { SiemPageName } from '../home/types';
+import { hostsModel } from '../../store';
+import { InputsModelId } from '../../store/inputs/constants';
+import { GlobalTimeArgs } from '../../containers/global_time';
+
+export const hostsPagePath = `/:pageName(${SiemPageName.hosts})`;
+export const hostDetailsPagePath = `${hostsPagePath}/:detailName`;
+
+export interface HostsComponentReduxProps {
+  query: Query;
+  filters: Filter[];
+}
+
+export interface HostsComponentDispatchProps {
+  setAbsoluteRangeDatePicker: ActionCreator<{
+    id: InputsModelId;
+    from: number;
+    to: number;
+  }>;
+  hostsPagePath: string;
+}
+
+export type HostsTabsProps = HostsComponentDispatchProps &
+  HostsQueryProps & {
+    filterQuery: string;
+    type: hostsModel.HostsType;
+    indexPattern: StaticIndexPattern;
+  };
+
+export type HostsQueryProps = GlobalTimeArgs;
+
+export type HostsComponentProps = HostsComponentReduxProps &
+  HostsComponentDispatchProps &
+  HostsQueryProps;
